requirements.txt
numpy
pandas
seaborn
matplotlib
pymysql
mysql-connector-python
python-dotenv
scikit-learn
xgboost
mlflow
imbalanced-learn
dill
#-e .